Buddy Console finished
+ other fixes and  return message changes

+++ GridWebServices/OrderService/Models/Order.cs
@@ -128,7 +128,7 @@ namespace OrderService.Models
         public int AccountID { get; set; }
         public string InvoiceName { get; set; }
         public string InvoiceUrl { get; set; }
-        public float FinalAmount { get; set; }
+        public double FinalAmount { get; set; }
         public string Remarks { get; set; }
         public int OrderStatus { get; set; }
         public int PaymentSourceID { get; set; }
@@ -143,6 +143,6 @@ namespace OrderService.Models
     {
         public string InvoiceID { get; set; }       
         public string InvoiceName { get; set; }       
-        public float FinalAmount { get; set; }
+        public double FinalAmount { get; set; }
     }
 }
